pyproject.toml
[project]
name = "scout"
version = "0.6.0"
description = "Browsing for Claude Code"
requires-python = ">=3.11"
authors = [{name = "sdoherty"}]
license = {text = "MIT"}
readme = "README.md"
keywords = ["browser", "automation", "anti-detection", "stealth", "mcp", "claude-code"]
dependencies = [
    "mcp[cli]>=1.0.0",
    "botasaurus-driver>=4.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[project.scripts]
scout = "scout.server:main"

[project.urls]
Homepage = "https://github.com/stemado/scout"
Repository = "https://github.com/stemado/scout"
Issues = "https://github.com/stemado/scout/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: tests requiring a real browser (deselect with -m 'not integration')",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.0.0",
]
